+++ operator/internal/resources/certificates/doc.go
@@ -1,0 +1,6 @@
+/*
+ * Copyright (c) 2026, NVIDIA CORPORATION. All rights reserved.
+ */
+
+// Package certificates contains shared cert-manager resource helpers.
+package certificates